Modified color behavior for remove/minus query box
If there is only 1 query box, the "minus" symbol will be black and the user won't be able to remove the query (the latter of which was added in [this](https://github.com/econpy/BookwormGUI/commit/f9cf431047667cb9cf2d6de7522d9a32af663bba) commit). But if there is more than 1 query box, the "minus" symbol will be red, designating the fact that a query box can be removed.